+++ bin/tprism/loss/base.py
@@ -5,8 +5,8 @@ from torch import Tensor
 import torch
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from tprism.expl_graph import SwitchTensorProvider
-    from tprism.torch_expl_graph import GoalInsideEntry
+    from tprism.expl_tensor import SwitchTensorProvider
+    from tprism.expl_graph import GoalInsideEntry
 
 class BaseLoss:
     def __init__(self, parameters: Optional[List[Any]]=None):
