+++ bin/tprism/loss/standard_loss.py
@@ -13,8 +13,8 @@ from torch import Tensor
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from tprism.expl_graph import SwitchTensorProvider
-    from tprism.torch_expl_graph import GoalInsideEntry
+    from tprism.expl_tensor import SwitchTensorProvider
+    from tprism.expl_graph import GoalInsideEntry
 
 class NLL(BaseLoss):
     def __init__(self, parameters: Optional[List[Any]]=None):
