+++ bin/tprism/op/standard_op.py
