+++ bin/tprism/plot/graph.py
@@ -1,6 +1,6 @@
 import networkx as nx
 from typing import Any, Dict, List, Optional, Set, Tuple
-from tprism.torch_expl_graph import GoalInsideEntry
+from tprism.expl_graph import GoalInsideEntry
 
 
 PathNode = Dict[str, Any]
